ref(client): assume app name is current directory
Previously, `deis create` would store session data in a ~/.deis/apps.yml
file for applications with no git root to facilitate `deis pull`. This
file looked similar to this:

    {'/tmp/deis-test': 'united-mainsail'}

The issue with this means if you move your application's root directory
somewhere else, the reference to the instance is lost. With this change,
if no git root is present in the application directory, the base name
of the current working directory (in the example above, that would be
'deis-test') is synonymous with the app name. With this in mind,
`deis create` in a git repo retains the old behaviour (via the
auto-generator from the server), but in a non-git directory, it takes
the name of the current working directory. As always, --app works just
like it used to, or you can still manually specify a name with
`deis create <id>`.